Implement check function for FritzBox smart home thermometers
This implements a new function to check the Fritz!Box DECT Smart Home
thermometers. The new function called `thermometer` will check the connection
status and the current temperature. Warning and Critical values can set for the
temperature. This also add a documentation about the index number that is needed
to check Smart Home devices and where you can get those.

refs #14